requirements.txt
praw==7.7.1
pandas==2.1.4
python-dotenv==1.0.0
openpyxl==3.1.2
requests==2.31.0
beautifulsoup4==4.12.2
textblob==0.17.1
nltk==3.8.1
schedule==1.2.0
tqdm==4.66.1
colorama==0.4.6
mistralai==1.9.11
